some x64 fixes

+++ src/g_main.c
@@ -96,7 +96,7 @@ This is the only way control passes into the module.
 This must be the very first function compiled into the .q3vm file
 ================
 */
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5,
+intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5,
 	    int arg6, int arg7, int arg8, int arg9, int arg10, int arg11 )
 {
     int api_ver;
@@ -121,7 +121,7 @@ int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int a
 		}
 
 		G_InitGame( arg0, arg1 );
-		return ( int ) ( &gamedata );
+		return ( intptr_t ) ( &gamedata );
 
 	case GAME_LOADENTS:
 		G_SpawnEntitiesFromString();
